add default ward
