--- packages/esm-ward-app/src/beds/unassigned-patient.scss
@@ -1,6 +1,0 @@
-@use '@openmrs/esm-styleguide/src/vars';
-
-.unassignedPatient {
-  display: flex;
-  flex-direction: column;
-}
