+++ packages/esm-ward-app/src/ward-patient-card/ward-patient-card.scss
@@ -188,3 +188,8 @@
 .skeletonText {
   width: 6.25rem;
 }
+
+.unassignedPatient {
+  display: flex;
+  flex-direction: column;
+}
