+++ packages/esm-ward-app/src/ward-view/default-ward/default-ward-unassigned-patients.component.tsx
@@ -1,0 +1,28 @@
+import { useAppContext } from '@openmrs/esm-framework';
+import React from 'react';
+import { type WardPatientGroupDetails } from '../../types';
+import DefaultWardPatientCard from '../../ward-patient-card/default-ward/default-ward-patient-card.component';
+
+function DefaultWardUnassignedPatients() {
+  const wardPatientsGrouping = useAppContext<WardPatientGroupDetails>('ward-patients-group');
+  const { wardUnassignedPatientsList } = wardPatientsGrouping ?? {};
+
+  const wardUnassignedPatients = wardUnassignedPatientsList?.map((inpatientAdmission) => {
+    return (
+      <DefaultWardPatientCard
+        {...{
+          patient: inpatientAdmission.patient,
+          visit: inpatientAdmission.visit,
+          bed: null,
+          inpatientAdmission,
+          inpatientRequest: inpatientAdmission.currentInpatientRequest,
+        }}
+        key={inpatientAdmission.patient.uuid}
+      />
+    );
+  });
+
+  return <>{wardUnassignedPatients}</>;
+}
+
+export default DefaultWardUnassignedPatients;
