+++ packages/esm-ward-app/src/ward-workspace/admission-request-card/admission-request-card.component.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import type { WardPatientCard } from '../../types';
+import type { WardPatientCardType } from '../../types';
 import AdmissionRequestCardActions from './admission-request-card-actions.component';
 import AdmissionRequestCardHeader from './admission-request-card-header.component';
 import styles from './admission-request-card.scss';
 
-const AdmissionRequestCard: WardPatientCard = (wardPatient) => {
+const AdmissionRequestCard: WardPatientCardType = (wardPatient) => {
   return (
     <div className={styles.admissionRequestCard}>
       <AdmissionRequestCardHeader {...wardPatient} />
